requirements.txt
fastapi
uvicorn
torch
torchvision
gradio
pillow
scikit-learn
numpy
pandas
tqdm
matplotlib
